requirements.txt
# ── Airflow (match the version in the Dockerfile FROM line) ──────────────────
apache-airflow==2.9.2
apache-airflow-providers-postgres==5.11.0

# ── PySpark ───────────────────────────────────────────────────────────────────
pyspark==3.5.1

# ── dbt ───────────────────────────────────────────────────────────────────────
dbt-core==1.8.4
dbt-postgres==1.8.2

# ── PostgreSQL JDBC driver (downloaded separately — see Dockerfile) ───────────
# psycopg2-binary is the Python Postgres adapter used by dbt + Airflow
psycopg2-binary==2.9.9

# ── Utilities ─────────────────────────────────────────────────────────────────
pandas==2.2.2
python-dotenv==1.0.1
